Encoding: extract uft8 coercion method from Xcrun
We need the coercion method in other places
