+++ lib/run_loop.rb
@@ -1,6 +1,7 @@
 require 'run_loop/regex'
 require 'run_loop/directory'
 require 'run_loop/environment'
+require "run_loop/encoding"
 require 'run_loop/logging'
 require 'run_loop/dot_dir'
 require 'run_loop/xcrun'
