+++ lib/run_loop/xcrun.rb
@@ -1,7 +1,8 @@
 module RunLoop
   class Xcrun
 
-    require 'command_runner'
+    require "command_runner"
+    include RunLoop::Encoding
 
     # Controls the behavior of Xcrun#exec.
     #
@@ -20,9 +21,6 @@ module RunLoop
 
     # Raised when Xcrun fails.
     class Error < RuntimeError; end
-
-    # Raised when the output of the command cannot be coerced to UTF8
-    class UTF8Error < RuntimeError; end
 
     # Raised when Xcrun times out.
     class TimeoutError < RuntimeError; end
@@ -60,7 +58,7 @@ IO.popen requires all arguments to be Strings.
         start_time = Time.now
         command_output = CommandRunner.run(['xcrun'] + args, timeout: timeout)
 
-        out = encode_utf8_or_raise(command_output[:out], cmd)
+        out = ensure_command_output_utf8(command_output[:out], cmd)
         process_status = command_output[:status]
 
         hash =
@@ -71,7 +69,7 @@ IO.popen requires all arguments to be Strings.
                     :exit_status => process_status.exitstatus
               }
 
-      rescue UTF8Error => e
+      rescue RunLoop::Encoding::UTF8Error => e
         raise e
       rescue => e
         elapsed = "%0.2f" % (Time.now - start_time)
@@ -98,35 +96,6 @@ with a timeout of #{timeout}
       end
 
       hash
-    end
-
-    private
-
-    # @!visibility private
-    def encode_utf8_or_raise(string, command)
-      return '' if !string
-
-      utf8 = string.force_encoding("UTF-8").chomp
-
-      return utf8 if utf8.valid_encoding?
-
-      encoded = utf8.encode('UTF-8', 'UTF-8', invalid: :replace, undef: :replace, replace: '')
-
-      return encoded if encoded.valid_encoding?
-
-        raise UTF8Error, %Q{
-Could not force UTF-8 encoding on this string:
-
-#{string}
-
-which is the output of this command:
-
-#{command}
-
-Please file an issue with a stacktrace and the text of this error.
-
-https://github.com/calabash/run_loop/issues
-}
     end
   end
 end
